Move log system error up to super
